requirements.txt
openai==1.91.0
litellm[proxy]

gradio==5.50.0

# Force CPU-only PyTorch wheels
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple
torch==2.5.1
transformers==4.45.2
sentence-transformers==3.3.1
networkx==3.4.2
pydantic==2.10.4
python_igraph==0.11.8
tenacity==8.5.0
tiktoken==0.7.0
nest_asyncio
python-dotenv==1.2.2
numpy
tqdm
pyarrow
neo4j==5.28.3
pymilvus==2.6.12